Refactored player hand to allow easier discard
The ui_player_hand was written when I was first starting on gamemaker. The original idea was that I wouldn't have to worry about empty space in the array so we could set the size at the beginning and we would fill it as we went. Now that I have a better understanding of the language this seems unnecessary as adding 1 index to the array will not take much work and this way simplifies a lot.