Made cards drawn from the player's current deck
I needed to make player_current_deck global. This is a little more dangerous because we have to make sure we empty it out any time we are entering a new encounter now. Hopefully adding to the initial_hand_setup is enough but we need to keep it in mind