Fixed buttons activating without click starting on it
The buttons used the mouse down events but that means it activates if you click and hold then mouse over the button. This is fixed by using the mouse_check_button_pressed. This seems to only activate once when the click happens, and as far as I can tell is not framerate dependent.